% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iiratehorton.R
\name{iiratehorton}
\alias{iiratehorton}
\title{Horton's empirical equation for instantaneous infiltration rate}
\usage{
iiratehorton(isubf, iinit, beta, t)
}
\arguments{
\item{isubf}{final infiltration rate @param iinit initial infiltration rate}

\item{beta}{coefficient describing the decay of the infiltration rate}

\item{t}{time}
}
\description{
This function allows you to calculate the instananeous infiltration rate
}
\examples{
iiratehorton() 
 
}
\keyword{empirical}
\keyword{infiltration}
\keyword{water}
