% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washoffFunction1.R
\name{washoffFunction1}
\alias{washoffFunction1}
\title{A typical washoff function}
\usage{
washoffFunction1(Pwinit, kw, r, t)
}
\arguments{
\item{Pwinit}{initial pollutant concentration @param kw empirical pollutant removal coefficient @param r runoff flowrate}
}
\description{
The empirical model for the mass of pollutant at time , t, given removal coefficient and runoff flow rate
}
\examples{
washoffFunction1() 
 
}
\keyword{empirical}
\keyword{quality}
\keyword{washoff}
\keyword{water}
